+++ pkgs/test/test/runner/pause_after_load_test.dart
@@ -7,7 +7,6 @@
 import 'dart:async';
 import 'dart:io';
 
-import 'package:pedantic/pedantic.dart';
 import 'package:test/test.dart';
 import 'package:test_descriptor/test_descriptor.dart' as d;